New RecordPage with successful ship order details page, removing the old broken record testing page
